usb: ci: add `disable-vbus-oc-protection` device-tree option
commit 49419a24b4c738b757524a06e6a5271bfb755094 from master-stream810

Some devices do not have the neccecary HW connections for the USB
over-current detection, in these cases we need to disable this feature
using a device-tree option.

[handle new workqueue struct]
[rebase to 6.6]

+++ drivers/usb/chipidea/core.c
@@ -841,6 +841,11 @@ static int ci_get_platdata(struct device *dev,
 	if (!platdata->enter_lpm)
 		platdata->enter_lpm = ci_hdrc_enter_lpm_common;
 
+	platdata->vbus_oc_protection_disabled = of_property_read_bool(dev->of_node, "disable-vbus-oc-protection");
+	if (platdata->vbus_oc_protection_disabled) {
+		dev_info(dev, "VBUS OC protetcion is disabled\n");
+	}
+
 	return 0;
 }
 
@@ -1320,18 +1325,20 @@ static int ci_hdrc_probe(struct platform_device *pdev)
 	hw_write_id_reg(ci, 0x230, (7 << 10), (0 << 10));
 
 	/* Init workqueue for checking VBUS overcurrent & start it. */
-	ci->vbus_oc_wq = create_freezable_workqueue("ci_vbus_oc");
-	if (!ci->vbus_oc_wq) {
-		dev_err(ci->dev, "can't create vbus_oc workqueue\n");
-		goto pm_stop;
-	}
 	ci->vbus_overcurrent = false;
-	INIT_DELAYED_WORK(&ci->check_vbus_work,
-			  imx8_check_vbus_overcurrent);
-	queue_delayed_work(
-		ci->vbus_oc_wq,
-		&ci->check_vbus_work,
-		msecs_to_jiffies(1000));
+	if (!ci->platdata->vbus_oc_protection_disabled) {
+		ci->vbus_oc_wq = create_freezable_workqueue("ci_vbus_oc");
+		if (!ci->vbus_oc_wq) {
+			dev_err(ci->dev, "can't create vbus_oc workqueue\n");
+			goto pm_stop;
+		}
+		INIT_DELAYED_WORK(&ci->check_vbus_work,
+				imx8_check_vbus_overcurrent);
+		queue_delayed_work(
+			ci->vbus_oc_wq,
+			&ci->check_vbus_work,
+			msecs_to_jiffies(1000));
+	}
 	return 0;
 
 pm_stop:
@@ -1368,8 +1375,10 @@ static void ci_hdrc_remove(struct platform_device *pdev)
 		pm_runtime_put_noidle(&pdev->dev);
 	}
 
-	flush_workqueue(ci->vbus_oc_wq);
-	destroy_workqueue(ci->vbus_oc_wq);
+	if (!ci->platdata->vbus_oc_protection_disabled) {
+		flush_workqueue(ci->vbus_oc_wq);
+		destroy_workqueue(ci->vbus_oc_wq);
+	}
 	dbg_remove_files(ci);
 	ci_role_destroy(ci);
 	ci_hdrc_enter_lpm(ci, true);
@@ -1475,10 +1484,11 @@ static int ci_controller_resume(struct device *dev)
 		ci_extcon_wakeup_int(ci);
 	}
 
-	queue_delayed_work(
-		ci->vbus_oc_wq,
-		&ci->check_vbus_work,
-		msecs_to_jiffies(0));
+	if (!ci->platdata->vbus_oc_protection_disabled)
+		queue_delayed_work(
+			ci->vbus_oc_wq,
+			&ci->check_vbus_work,
+			msecs_to_jiffies(0));
 
 	return 0;
 }
@@ -1488,7 +1498,8 @@ static int ci_suspend(struct device *dev)
 {
 	struct ci_hdrc *ci = dev_get_drvdata(dev);
 
-	flush_workqueue(ci->vbus_oc_wq);
+	if (!ci->platdata->vbus_oc_protection_disabled)
+		flush_workqueue(ci->vbus_oc_wq);
 	if (ci->wq)
 		flush_workqueue(ci->wq);
 	/*
@@ -1575,7 +1586,8 @@ static int ci_runtime_suspend(struct device *dev)
 	if(ci->vbus_overcurrent)
 		return -EBUSY;
 
-	cancel_delayed_work_sync(&ci->check_vbus_work); 
+	if (!ci->platdata->vbus_oc_protection_disabled)
+		cancel_delayed_work_sync(&ci->check_vbus_work); 
 
 	if (ci->in_lpm)
 		return 0;
